chore(cms): move rbacStore to server namespace
