+++ apps/cms/src/lib/server/rbacStore.ts
@@ -1,4 +1,6 @@
-// apps/cms/src/lib/rbacStore.ts
+import "server-only";
+
+// apps/cms/src/lib/server/rbacStore.ts
 
 import type { CmsUser } from "@acme/types";
 import type { Permission } from "@auth";
@@ -6,7 +8,7 @@ import { ROLE_PERMISSIONS } from "@auth/permissions";
 import * as fsSync from "fs";
 import { promises as fs } from "fs";
 import * as path from "path";
-import type { Role } from "../auth/roles";
+import type { Role } from "../../auth/roles";
 
 export interface RbacDB {
   users: Record<string, CmsUser>;
